requirements.txt
Django
mysql
gunicorn
mysqlclient
python-dotenv
openai
PyPDF2
opencv-python
numpy
pymysql
cryptography
mysql-connector-python
